@@ -1,62 +1,64 @@
-name: syntactic
+name: nahawi
 description: "A new Flutter project."
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 
-version: 1.0.0+1
+version: 1.0.5+5
 
 environment:
-  sdk: '>=3.2.3 <4.0.0'
+  sdk: '>=3.3.4 <4.0.0'
 
 
 dependencies:
   flutter:
     sdk: flutter
 
-
-
-  cupertino_icons: ^1.0.6
-  get_it: ^7.6.4
+  cupertino_icons: ^1.0.8
+  get_it: ^7.7.0
   shared_preferences: ^2.2.2
   get: ^4.6.6
-  animated_theme_switcher: ^2.0.9
+  animated_theme_switcher: ^2.0.10
   another_xlider: ^3.0.2
-  flutter_svg: ^2.0.9
+  flutter_svg: ^2.0.10+1
   gap: ^3.0.1
   hijri: ^3.0.0
   flutter_slider_drawer: ^2.1.3
   arabic_numbers: ^1.0.0
-  screenshot: ^2.1.0
-  path_provider: ^2.1.1
-  share_plus: ^7.2.1
-  flutter_screenutil: ^5.9.0
-  stylish_bottom_bar: ^1.0.3
+  screenshot: ^3.0.0
+  path_provider: ^2.1.4
+  share_plus: ^10.0.0
+  flutter_screenutil: ^5.9.3
+  stylish_bottom_bar: ^1.1.0
   expansion_tile_card: ^3.0.0
-  device_info_plus: ^9.1.1
-  http: ^1.1.2
-  url_launcher: ^6.2.2
+  device_info_plus: ^10.1.1
+  http: ^1.2.2
+  url_launcher: ^6.3.0
   info_popup: ^4.3.1
-  lottie: ^2.7.0
-  pie_menu: ^3.1.3
-  just_audio: ^0.9.36
-  connectivity_plus: ^5.0.2
-  dio: ^5.4.0
-  objectbox: ^2.4.0
+  lottie: ^3.1.2
+  pie_menu: ^3.2.1
+  just_audio: ^0.9.39
+  connectivity_plus: ^6.0.4
+  dio: ^5.5.0+1
+  objectbox: ^4.0.1
   objectbox_flutter_libs: any
   diacritic: ^0.1.5
-  group_button: ^5.3.3
-  image: ^4.1.3
+  group_button: ^5.3.4
+  image: ^4.2.0
   horizontal_stepper_step: ^0.0.5
-  flutter_native_splash: ^2.3.9
+  flutter_native_splash: ^2.3.13
+  path: ^1.9.0
+  package_rename: ^1.7.1
+  rxdart: ^0.28.0
+  easy_url_launcher: ^0.0.7
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
-  flutter_lints: ^3.0.1
-  build_runner: ^2.4.7
+  flutter_lints: ^4.0.0
+  build_runner: ^2.4.8
   objectbox_generator: any
 
 
@@ -72,6 +74,8 @@ flutter_native_splash:
     image_dark: assets/images/syntactic_logo240.png
     icon_background_color_dark: "#181818"
 
+
+
 flutter:
 
 
@@ -85,6 +89,7 @@ flutter:
     - assets/icons/
     - assets/lottie/
     - assets/images/onboarding/
+    - assets/images/
     - assets/json/bookName.json
     - assets/json/poems/0.json
     - assets/json/poems/1.json
