@@ -25,9 +25,9 @@ List<String> imagesD = <String>[
 ];
 
 List<String> onboardingTitle = <String>[
-  'السلام عليكم، حياكم الله وبياكم في تطبيق "نحويّ"',
+  'السلام عليكم، حياكم الله وبياكم في تطبيق "نحويّ : النحو العربي"',
   'واجهة سهلة الاستخدام',
   'إضافة شارات مرجعية والرجوع اليها في أي وقت آخر',
-  'مشاركة المحتوى كنص أو كصورة',
+  'البحث في كل الكتب',
   'الاستماع لمحتوى بعض الكتب',
 ];
